Added order routes

+++ controllers/blogController.js
@@ -6,7 +6,7 @@ const multer = require('multer');
 
 const Blog = require('../models/blogModel');
 const Contents = require('../models/contentsModel');
-const { randomId } = require('../utils/helpers');
+const { randomId, fileNameInKebabCase } = require('../utils/helpers');
 
 const multerStorage = multer.memoryStorage();
 
@@ -25,11 +25,12 @@ const upload = multer({
 
 exports.uploadBlogImage = upload.single('cover');
 
+// CREATE BLOG
 exports.createBlog = async (req, res, next) => {
   try {
     const { title, contents } = req.body;
 
-    console.log(title, JSON.parse(contents));
+    // console.log(title, JSON.parse(contents));
     // console.log(req.file);
 
     if (!title || !contents) {
@@ -48,17 +49,8 @@ exports.createBlog = async (req, res, next) => {
       mkdirSync(dir, { recursive: true });
     }
 
-    let id = randomId();
-    const fileName =
-      req.file.originalname
-        .toLowerCase()
-        .replaceAll(' ', '-')
-        .replace(/\s/g, '-')
-        .split('.')[0] +
-      '-' +
-      id +
-      '.' +
-      req.file.originalname.split('.')[1];
+    // let id = randomId();
+    const fileName = fileNameInKebabCase(req.file.originalname);
 
     await writeFile(join(dir, fileName), req.file.buffer);
 
@@ -71,14 +63,6 @@ exports.createBlog = async (req, res, next) => {
       cover: blogCover,
     });
 
-    /*
-    [
-  { title: 'title2', description: 'desc 1' },
-  { title: 'title2', description: 'desc 2' },
-  { title: 'title3', description: 'desc3' }
-]
-    */
-
     const newContents = await Contents.insertMany(
       JSON.parse(contents).map((content) => ({
         blogId: newblog._id,
@@ -88,7 +72,7 @@ exports.createBlog = async (req, res, next) => {
     );
 
     // Update the blog with the contents
-    const update = await Blog.findOneAndUpdate(
+    const updateBlogWithContents = await Blog.findOneAndUpdate(
       {
         _id: newblog._id,
       },
@@ -100,32 +84,22 @@ exports.createBlog = async (req, res, next) => {
         // runValidators: true,
       }
     );
-    console.log('update', update);
+    // console.log('update', updateBlogWithContents);
 
     // Reset the id
-    id = '';
+    // id = '';
 
-    return res.status(201).json({ newblog });
-  } catch (error) {
-    console.error(error.stack);
-    return res.status(500).json({ message: error.message });
-  }
-};
-
-exports.getContents = async (req, res, next) => {
-  try {
-    const contents = await Contents.find({}).lean().exec();
-
-    if (!contents) {
-      return res.status(404).json({ message: 'success', contents });
-    }
-
-    return res.status(200).json({ contents });
+    return res.status(201).json({
+      status: 'success',
+      message: 'Successfully created blog post.',
+      data: updateBlogWithContents,
+    });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 };
 
+// GET BLOGS
 exports.getBlogs = async (req, res, next) => {
   try {
     const blogs = await Blog.find({})
@@ -144,59 +118,130 @@ exports.getBlogs = async (req, res, next) => {
   }
 };
 
+// GET BLOG
 exports.getBlog = async (req, res, next) => {
   try {
     const { id } = req.params;
 
-    const blog = await Blog.findById({ _id: id });
+    if (!id || id <= 24) {
+      return res.status(400).json({ message: 'Please provide a valid id.' });
+    }
+
+    const blog = await Blog.findById({ _id: id })
+      .lean()
+      .select('-updatedAt')
+      .populate('contents', 'title description')
+      .exec();
 
     if (!blog) {
       return res.status(404).json({ message: 'Blog not found' });
     }
 
-    return res.status(200).json({ blog });
+    return res.status(200).json({ status: 'success', blog });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 };
 
+// UPDATE BLOG
 exports.updateBlog = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const { title, content } = req.body;
+    const { title } = req.body;
+    // const { cover } = req.file;
 
-    if (!title || !content) {
-      return res
-        .status(400)
-        .json({ message: 'Please provide all the details' });
+    const serverUrl = `https://weroambags-backend.onrender.com`;
+    let fileName = undefined;
+
+    // If there no id or is less than 24 characters
+    if (!id || id <= 24) {
+      return res.status(400).json({ message: 'Please provide a valid id.' });
     }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(
+    // If there is no title or content
+    if (!title) {
+      return res.status(400).json({ message: 'Please provide the title.' });
+    }
+
+    // find the existing blog
+    const existingBlog = await Blog.findById({ _id: id }).lean();
+    if (!existingBlog) {
+      return res.status(404).json({ message: 'No Blog found for update.' });
+    }
+
+    // if there is a new cover image
+    if (req.file) {
+      // delete the existing blog cover
+      const existingCover = existingBlog.cover.split('/').pop();
+      const filePath = join(__dirname, '..', 'public', 'blogs', existingCover);
+
+      // find the existing blog cover
+      if (existsSync(filePath)) {
+        await unlink(filePath);
+      }
+
+      const dir = join(__dirname, '..', 'public', 'blogs');
+      fileName = fileNameInKebabCase(req.file.originalname);
+
+      await writeFile(join(dir, fileName), req.file.buffer);
+    }
+
+    const updateObj = {
+      title: title || existingBlog.title,
+      cover: fileName ? `${serverUrl}/blogs/${fileName}` : existingBlog.cover,
+    };
+
+    const updatedBlog = await Blog.findOneAndUpdate(
       {
         _id: id,
       },
       {
-        $set: { title, content },
+        $set: updateObj,
       },
       {
         new: true,
-        runValidators: true,
+        // runValidators: true,
       }
     );
 
     if (!updatedBlog) {
-      return res.status(404).json({ message: 'Blog not found' });
+      return res.status(400).json({ message: 'Something went wrong' });
     }
 
-    return res.status(200).json({ updatedBlog });
+    // Reset the fileName
+    fileName = undefined;
+
+    return res
+      .status(200)
+      .json({ status: 'Successfully updated blog.', updatedBlog });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 };
 
+// DELETE BLOG
 exports.deleteBlog = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!id || id <= 24) {
+      return res.status(400).json({ message: 'Please provide a valid id.' });
+    }
+
+    // find the existing blog
+    const blog = await Blog.findById({ _id: id }).lean();
+
+    if (!blog) {
+      return res.status(404).json({ message: 'Blog not found for delete.' });
+    }
+
+    // delete the blog cover
+    const cover = blog.cover.split('/').pop();
+    const dir = join(__dirname, '..', 'public', 'blogs', cover);
+
+    if (existsSync(dir)) {
+      await unlink(dir);
+    }
 
     const deletedBlog = await Blog.findByIdAndDelete(
       { _id: id },
@@ -206,22 +251,45 @@ exports.deleteBlog = async (req, res, next) => {
       }
     );
 
-    if (!deletedBlog) {
-      return res.status(404).json({ message: 'Blog not found' });
+    // delete the contents
+    const deletedContents = await Contents.deleteMany({ blogId: id });
+
+    if (!deletedBlog || !deletedContents) {
+      return res.status(404).json({ message: 'Something went wrong' });
     }
 
     return res.status(200).json({
       status: 'success',
-      message: 'Successfully deleted the blog post',
+      message: 'Successfully deleted the blog post and its contents',
     });
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 };
 
+// GET CONTENTS
+exports.getContents = async (req, res, next) => {
+  try {
+    const contents = await Contents.find({}).lean().exec();
+
+    if (!contents) {
+      return res.status(404).json({ message: 'success', contents });
+    }
+
+    return res.status(200).json({ contents });
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
+  }
+};
+
+// DELETE CONTENT
 exports.deleteContent = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!id || id <= 24) {
+      return res.status(400).json({ message: 'Please provide a valid id.' });
+    }
 
     const deletedContent = await Contents.findByIdAndDelete(
       { _id: id },
